move helper functions into another files
