+++ services/frontend/package.json
@@ -9,7 +9,7 @@
     "build": "NODE_ENV=production craco build",
     "test": "craco test",
     "lint": "./node_modules/.bin/eslint --cache --format codeframe --ext jsx,js src",
-    "i18n:check": "node ./src/i18n/check",
+    "i18n:check": "node ./src/i18n/validation",
     "i18n:cleanup": "node ./src/i18n/cleanUpUnusedKeys",
     "analyze": "ANALYZE_BUILD=true yarn build"
   },
